+++ RELEASE_NOTES.md
@@ -92,6 +92,7 @@ Note that the library no longer sets up English services by default, so make sur
 * `FeatureToggle` has been removed.
 * `HapticFeedback.Engine` is no longer open to inheritance.
 * `KeyboardAction.emojiCategory` has been removed.
+* `KeyboardCharacterProvider` has been made removed. Use `String` extensions directly instead.
 * `KeyboardColor` has been made internal.
 * `KeyboardColorReader` has been made removed. Use `Color` extensions directly instead.
 * `KeyboardSettingsUrlProvider` has been removed. Use `URL.keyboardSettings` directly instead.